code/bg-remove/requirements.txt
flask==2.0.1
rembg==2.0.50
Pillow==9.5.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.13.0
gunicorn==20.1.0